refactor: move tolerant zip handling into brut.j.zip
Per review: introduce a dedicated brut.j.zip module owning the tolerant
zip behavior, and reduce the change outside it to swapping the backing
ZipFile in brut.j.dir. ApkDecoder and ZipUtils return to their original
shape; decode routes through ExtFile -> ZipRODirectory as before.

RepairingZipFile repairs central-directory / local-header fields AOSP
tolerates (spurious encrypted bit, non-DEFLATE/STORE methods) in a temp
copy and removes it on close. Clean archives open in place.

The old apktool-lib test needed the full testapp apparatus; the module
test builds a minimal zip and flips the header bits directly.

Refs #4028
