+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/ApkDecoder.java
@@ -29,7 +29,6 @@ import brut.directory.ExtFile;
 import brut.util.BackgroundWorker;
 import brut.util.BrutIO;
 import brut.util.OS;
-import brut.util.ZipUtils;
 import org.apache.commons.io.FilenameUtils;
 
 import java.io.*;
@@ -47,8 +46,6 @@ public class ApkDecoder {
       + "rtttl|imy|xmf|mp4|m4a|m4v|3gp|3gpp|3g2|3gpp2|amr|awb|wma|wmv|webm|webp|mkv");
 
     private final ExtFile mApkFile;
-    private final ExtFile mZipSource;
-    private final File mRepairCopy;
     private final Config mConfig;
     private final AtomicReference<AndrolibException> mFirstError;
 
@@ -59,19 +56,6 @@ public class ApkDecoder {
 
     public ApkDecoder(File apkFile, Config config) {
         mApkFile = new ExtFile(apkFile);
-        File zipPath = apkFile;
-        File repairCopy = null;
-        try {
-            File opened = ZipUtils.openReadableZip(apkFile);
-            if (!opened.equals(apkFile)) {
-                repairCopy = opened;
-                zipPath = opened;
-            }
-        } catch (IOException ignored) {
-            // Unreadable archives fail later with the underlying zip error.
-        }
-        mZipSource = new ExtFile(zipPath);
-        mRepairCopy = repairCopy;
         mConfig = config;
         mFirstError = new AtomicReference<>();
     }
@@ -102,9 +86,8 @@ public class ApkDecoder {
         try {
             mApkInfo = new ApkInfo();
             mApkInfo.setVersion(mConfig.getVersion());
-            mApkInfo.setApkFile(mZipSource);
-            mApkInfo.setApkFileName(mApkFile.getName());
-            mSmaliDecoder = new SmaliDecoder(mZipSource, mConfig.isBaksmaliDebugMode());
+            mApkInfo.setApkFile(mApkFile);
+            mSmaliDecoder = new SmaliDecoder(mApkFile, mConfig.isBaksmaliDebugMode());
             mResDecoder = new ResDecoder(mApkInfo, mConfig);
 
             Log.i(TAG, "Using Apktool " + mConfig.getVersion() + " on " + mApkFile.getName()
@@ -130,11 +113,8 @@ public class ApkDecoder {
                 mWorker.shutdownNow();
             }
             try {
-                mZipSource.close();
+                mApkFile.close();
             } catch (DirectoryException ignored) {
-            }
-            if (mRepairCopy != null) {
-                OS.rmfile(mRepairCopy);
             }
         }
     }
@@ -149,7 +129,7 @@ public class ApkDecoder {
         }
 
         try {
-            Directory in = mZipSource.getDirectory();
+            Directory in = mApkFile.getDirectory();
             boolean allSrc = mConfig.isDecodeSourcesFull();
             boolean noSrc = mConfig.isDecodeSourcesNone();
 
@@ -172,7 +152,7 @@ public class ApkDecoder {
     private void copySourcesRaw(File outDir, String fileName) throws AndrolibException {
         Log.i(TAG, "Copying raw " + fileName + "...");
         try {
-            Directory in = mZipSource.getDirectory();
+            Directory in = mApkFile.getDirectory();
 
             in.copyToDir(outDir, fileName);
         } catch (DirectoryException ex) {
@@ -216,7 +196,7 @@ public class ApkDecoder {
     private void copyResourcesRaw(File outDir) throws AndrolibException {
         Log.i(TAG, "Copying raw resources.arsc...");
         try {
-            Directory in = mZipSource.getDirectory();
+            Directory in = mApkFile.getDirectory();
 
             in.copyToDir(outDir, "resources.arsc");
         } catch (DirectoryException ex) {
@@ -239,7 +219,7 @@ public class ApkDecoder {
     private void copyManifestRaw(File outDir) throws AndrolibException {
         Log.i(TAG, "Copying raw AndroidManifest.xml...");
         try {
-            Directory in = mZipSource.getDirectory();
+            Directory in = mApkFile.getDirectory();
 
             in.copyToDir(outDir, "AndroidManifest.xml");
         } catch (DirectoryException ex) {
@@ -249,7 +229,7 @@ public class ApkDecoder {
 
     private void copyRawFiles(File outDir) throws AndrolibException {
         try {
-            Directory in = mZipSource.getDirectory();
+            Directory in = mApkFile.getDirectory();
             Set<String> dexFiles = mSmaliDecoder.getDexFiles();
             Map<String, String> resFileMap = mResDecoder.getResFileMap();
             boolean noAssets = mConfig.isDecodeAssetsNone();
@@ -278,7 +258,7 @@ public class ApkDecoder {
 
         Log.i(TAG, "Copying original files...");
         try {
-            Directory in = mZipSource.getDirectory();
+            Directory in = mApkFile.getDirectory();
 
             for (String fileName : in.getFiles(true)) {
                 if (ApkInfo.ORIGINAL_FILES_PATTERN.matcher(fileName).matches()) {
@@ -295,7 +275,7 @@ public class ApkDecoder {
 
         Log.i(TAG, "Copying unknown files...");
         try {
-            Directory in = mZipSource.getDirectory();
+            Directory in = mApkFile.getDirectory();
             Set<String> dexFiles = mSmaliDecoder.getDexFiles();
             Map<String, String> resFileMap = mResDecoder.getResFileMap();
 
@@ -321,7 +301,7 @@ public class ApkDecoder {
 
         // Record uncompressed files.
         try {
-            Directory in = mZipSource.getDirectory();
+            Directory in = mApkFile.getDirectory();
             Map<String, String> resFileMap = mResDecoder.getResFileMap();
             Set<String> uncompressedExts = new HashSet<>();
             Set<String> uncompressedFiles = new HashSet<>();
