+++ brut.j.dir/build.gradle.kts
@@ -1,4 +1,5 @@
 dependencies {
     implementation(project(":brut.j.common"))
     implementation(project(":brut.j.util"))
+    implementation(project(":brut.j.zip"))
 }
