+++ brut.j.util/src/main/java/brut/util/ZipUtils.java
@@ -23,182 +23,19 @@ import org.apache.commons.io.IOUtils;
 import java.io.File;
 import java.io.InputStream;
 import java.io.IOException;
-import java.io.RandomAccessFile;
 import java.nio.file.Files;
 import java.nio.file.InvalidPathException;
-import java.nio.file.StandardCopyOption;
 import java.util.Collection;
 import java.util.function.Predicate;
 import java.util.zip.CRC32;
 import java.util.zip.ZipEntry;
-import java.util.zip.ZipException;
-import java.util.zip.ZipFile;
 import java.util.zip.ZipOutputStream;
 
 public final class ZipUtils {
     private static final String TAG = "";
 
-    private static final int EOCD_SIG = 0x06054b50;
-    private static final int CDFH_SIG = 0x02014b50;
-    private static final int LFH_SIG = 0x04034b50;
-    private static final int CDFH_LEN = 46;
-    private static final int EOCD_MIN_LEN = 22;
-    private static final int COMMENT_MAX_LEN = 0xFFFF;
-    private static final int EOCD_MAX_SEARCH = EOCD_MIN_LEN + COMMENT_MAX_LEN;
-
     private ZipUtils() {
         // Private constructor for utility class.
-    }
-
-    /**
-     * Copies {@code source} and repairs central-directory / local-header fields that AOSP tolerates
-     * but {@code java.util.zip.ZipFile} rejects (spurious encrypted bit, non-DEFLATE/STORE methods).
-     */
-    public static void ensureZipSystemProperties() {
-        // Match brut.apktool.Main — needed when ApkDecoder is used without the CLI wrapper.
-        System.setProperty("jdk.util.zip.disableZip64ExtraFieldValidation", "true");
-    }
-
-    public static File openReadableZip(File source) throws IOException {
-        ensureZipSystemProperties();
-        try (ZipFile zip = new ZipFile(source)) {
-            return source;
-        } catch (ZipException ex) {
-            if (!isRepairableCenHeader(ex)) {
-                throw ex;
-            }
-            return repairInvalidCenHeadersCopy(source);
-        }
-    }
-
-    public static boolean isRepairableCenHeader(IOException ex) {
-        if (ex instanceof ZipException) {
-            String message = ex.getMessage();
-            return message != null && message.contains("invalid CEN header");
-        }
-        return false;
-    }
-
-    public static File repairInvalidCenHeadersCopy(File source) throws IOException {
-        File dest = Files.createTempFile("apktool-zip-repair-", ".zip").toFile();
-        dest.deleteOnExit();
-        Files.copy(source.toPath(), dest.toPath(), StandardCopyOption.REPLACE_EXISTING);
-        repairInvalidCenHeaders(dest);
-        return dest;
-    }
-
-    public static void repairInvalidCenHeaders(File file) throws IOException {
-        try (RandomAccessFile f = new RandomAccessFile(file, "rw")) {
-            byte[] eocdData = findEocd(f);
-            int cdfhCnt = readUShort(eocdData, 10);
-            int cdOff = readUInt(eocdData, 16);
-            f.seek(cdOff);
-            int cdfhIdx = 0;
-            while (cdfhIdx < cdfhCnt) {
-                long cdfhOff = f.getFilePointer();
-                long cdfhSig = readUInt(f);
-                if (cdfhSig != CDFH_SIG) {
-                    throw new IOException("invalid central directory file header signature");
-                }
-                f.seek(cdfhOff + 8);
-                fixHeaderFields(f);
-                f.seek(cdfhOff + 28);
-                int nameLen = readUShort(f);
-                int extraLen = readUShort(f);
-                int commentLen = readUShort(f);
-                f.seek(cdfhOff + 42);
-                long lfhOff = readUInt(f);
-                f.seek(lfhOff);
-                long lfhSig = readUInt(f);
-                if (lfhSig != LFH_SIG) {
-                    throw new IOException("invalid local file header signature");
-                }
-                f.seek(lfhOff + 6);
-                fixHeaderFields(f);
-                f.seek(cdfhOff + CDFH_LEN + nameLen + extraLen + commentLen);
-                cdfhIdx++;
-            }
-        }
-    }
-
-    private static byte[] findEocd(RandomAccessFile f) throws IOException {
-        long size = f.length();
-        int maxSearch = (int) Math.min(size, EOCD_MAX_SEARCH);
-        f.seek(size - maxSearch);
-        byte[] data = new byte[maxSearch];
-        f.readFully(data);
-        int idx = lastIndexOf(data, EOCD_SIG);
-        if (idx < 0) {
-            throw new IOException("end of central directory record not found");
-        }
-        byte[] eocd = new byte[EOCD_MIN_LEN];
-        System.arraycopy(data, idx, eocd, 0, EOCD_MIN_LEN);
-        return eocd;
-    }
-
-    private static void fixHeaderFields(RandomAccessFile f) throws IOException {
-        long flagsPos = f.getFilePointer();
-        int flags = readUShort(f);
-        if ((flags & 0x1) != 0) {
-            f.seek(flagsPos);
-            writeUShort(f, flags & ~0x1);
-        }
-        long comprPos = flagsPos + 2;
-        f.seek(comprPos);
-        int compr = readUShort(f);
-        if (compr != 0 && compr != 8) {
-            f.seek(comprPos);
-            writeUShort(f, 0);
-        }
-    }
-
-    private static int lastIndexOf(byte[] data, int value) {
-        int b0 = value & 0xff;
-        int b1 = (value >> 8) & 0xff;
-        int b2 = (value >> 16) & 0xff;
-        int b3 = (value >> 24) & 0xff;
-        for (int i = data.length - 4; i >= 0; i--) {
-            if (data[i] == b0 && data[i + 1] == b1 && data[i + 2] == b2 && data[i + 3] == b3) {
-                return i;
-            }
-        }
-        return -1;
-    }
-
-    private static int readUShort(RandomAccessFile f) throws IOException {
-        int b0 = f.read();
-        int b1 = f.read();
-        if (b0 < 0 || b1 < 0) {
-            throw new IOException("unexpected end of file");
-        }
-        return (b1 << 8) | b0;
-    }
-
-    private static long readUInt(RandomAccessFile f) throws IOException {
-        long b0 = f.read();
-        long b1 = f.read();
-        long b2 = f.read();
-        long b3 = f.read();
-        if (b0 < 0 || b1 < 0 || b2 < 0 || b3 < 0) {
-            throw new IOException("unexpected end of file");
-        }
-        return b0 | (b1 << 8) | (b2 << 16) | (b3 << 24);
-    }
-
-    private static void writeUShort(RandomAccessFile f, int value) throws IOException {
-        f.write(value & 0xff);
-        f.write((value >> 8) & 0xff);
-    }
-
-    private static int readUShort(byte[] data, int offset) {
-        return (data[offset + 1] << 8) | (data[offset] & 0xff);
-    }
-
-    private static int readUInt(byte[] data, int offset) {
-        return (data[offset] & 0xff)
-            | ((data[offset + 1] & 0xff) << 8)
-            | ((data[offset + 2] & 0xff) << 16)
-            | ((data[offset + 3] & 0xff) << 24);
     }
 
     public static void zipDir(File dir, ZipOutputStream out, Collection<String> doNotCompress) throws IOException {
