+++ settings.gradle.kts
@@ -5,7 +5,7 @@ plugins {
 
 rootProject.name = "apktool-cli"
 include(
-    "brut.j.common", "brut.j.util", "brut.j.dir", "brut.j.xml", "brut.j.yaml",
+    "brut.j.common", "brut.j.util", "brut.j.dir", "brut.j.xml", "brut.j.yaml", "brut.j.zip",
     "brut.apktool:apktool-lib", "brut.apktool:apktool-cli"
 )
